feat: Implement animations using framer-motion for various components and update service cards
