+++ src/App.jsx
@@ -3,12 +3,22 @@ import Footer from "./pages/Footer";
 import Navbar from "./pages/Navbar";
 import Home from "./pages/Hero/Index";
 import { Toaster } from "react-hot-toast";
+import { motion } from "framer-motion";
 
 // Lazy load components
 const About = lazy(() => import("./pages/About/About"));
 const Project = lazy(() => import("./pages/Project"));
 const Contact = lazy(() => import("./pages/Contact"));
 const Service = lazy(() => import("./pages/Service"));
+
+const animation = {
+  section: {
+    initial: { opacity: 0, y: 50 },
+    whileInView: { opacity: 1, y: 0 },
+    transition: { duration: 0.6 },
+    viewport: { once: true, margin: "-100px" },
+  },
+};
 
 // Custom Error Boundary Component
 class ErrorBoundary extends React.Component {
@@ -65,31 +75,63 @@ function App() {
               </Suspense>
 
               <Suspense fallback={<LoadingFallback />}>
-                <section id="about">
+                <motion.section
+                  initial={animation.section.initial}
+                  whileInView={animation.section.whileInView}
+                  transition={animation.section.transition}
+                  viewport={animation.section.viewport}
+                  id="about"
+                >
                   <About />
-                </section>
+                </motion.section>
               </Suspense>
 
               <Suspense fallback={<LoadingFallback />}>
-                <section id="service">
+                <motion.section
+                  initial={animation.section.initial}
+                  whileInView={animation.section.whileInView}
+                  transition={animation.section.transition}
+                  viewport={animation.section.viewport}
+                  id="service"
+                >
                   <Service />
-                </section>
-              </Suspense>
-              <Suspense fallback={<LoadingFallback />}>
-                <section id="projects">
-                  <Project />
-                </section>
+                </motion.section>
               </Suspense>
 
               <Suspense fallback={<LoadingFallback />}>
-                <section id="contact">
-                  <Contact />
-                </section>
+                <motion.section
+                  initial={animation.section.initial}
+                  whileInView={animation.section.whileInView}
+                  transition={animation.section.transition}
+                  viewport={animation.section.viewport}
+                  id="projects"
+                >
+                  <Project />
+                </motion.section>
               </Suspense>
+
               <Suspense fallback={<LoadingFallback />}>
-                <section id="footer">
+                <motion.section
+                  initial={animation.section.initial}
+                  whileInView={animation.section.whileInView}
+                  transition={animation.section.transition}
+                  viewport={animation.section.viewport}
+                  id="contact"
+                >
+                  <Contact />
+                </motion.section>
+              </Suspense>
+
+              <Suspense fallback={<LoadingFallback />}>
+                <motion.section
+                  initial={animation.section.initial}
+                  whileInView={animation.section.whileInView}
+                  transition={animation.section.transition}
+                  viewport={animation.section.viewport}
+                  id="footer"
+                >
                   <Footer />
-                </section>
+                </motion.section>
               </Suspense>
             </main>
           </>
